\name{Asthma3}
\alias{Asthma3}
\docType{data}
\title{
Effect of curcuma longa on de-granulation of mast cells in mice
}
\description{
Mast cells if de-granulated release histamine causing allergic reaction. It is of interest to reduce de-granulation percentage.
}
\usage{data(Asthma3)}
\format{
  A data frame with 15 observations on the following 5 variables.
  \describe{
    \item{\code{Treatment}}{Treatment types}
    \item{\code{Animal.Code}}{The Animal Code}
    \item{\code{Response}}{Response as a percentage of the de-granulated cells}
  }
}
\details{
ANOVA, multiple comparisons, transformation methods are recommended to be performed on the data. }
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\examples{
data(Asthma3)
}
\keyword{datasets}
