\name{Bamboolife}
\alias{Bamboolife}
\docType{data}
\title{
Preparing a life table for the Bamboo plant
}
\description{
Bamboo is a peculiar plant. Individual plant is a cluster (called clump) of shoots called culms. A clump has a long life of about 40 years and produces seeds in super abundance only once, at the end of life. Further all bamboo clumps in a region develop flowers and seeds together and die together at the same time. There is no variation in age at death.  However, there is variation in the lifespan of individual culms. Data is on survival of 439 culms of bamboo (Dendrocalamus strictus). }
\usage{data(Bamboolife)}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{Age in years}
    \item{\code{Survivors}}{Number of survivors}
  }
}
\details{
Life tables is suggested for the current data set. }
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\examples{
data(Bamboolife)
}
\keyword{datasets}
