\name{BirthDeath}
\alias{BirthDeath}
\docType{data}
\title{
Changes in Human birth and death rates in India over the 20th century 
}
\description{
Birth and death rates are important indices of demographic picture of a country. Given data for last century can we predict the rates and net population growth for next decade?
}
\usage{data(BirthDeath)}
\format{
  A data frame with 27 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{The year grip}
    \item{\code{Birth.rate}}{The birth rate}
    \item{\code{death.rate}}{The death rate}
  }
}
\details{
Analyze the time trends in birth and death rates. Time trends in net growth rate.
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\examples{
data(BirthDeath)
}
\keyword{datasets}
