\name{Chitalparasite}
\alias{Chitalparasite}
\docType{data}
\title{
Understanding the correlation of occurrence of a parasite
}
\description{
It is of interest to examine relation between presence/ absence of parasite and other attributes. 
}
\usage{data(Chitalparasite)}
\format{
  A data frame with 66 observations on the following 10 variables.
  \describe{
    \item{\code{Sarcocystis_Indicator}}{a numeric vector}
    \item{\code{Sanctuary_Indicator}}{a numeric vector}
    \item{\code{Predator_Indicator}}{a numeric vector}
    \item{\code{Tissue_Indicator}}{a numeric vector}
    \item{\code{SEX}}{Gender}
    \item{\code{YEAR}}{Year}
  }
}
\details{
Explore contingency tables and logistic regression model for this data set. }
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\examples{
data(Chitalparasite)
}
\keyword{datasets}
