\name{Cosmetic1}
\alias{Cosmetic1}
\docType{data}
\title{
Testing efficacy of a cosmetic product 
}
\description{
A manufacturer of cosmetic products is interested in frequent introduction of new formulations in to the market. This may be for one of two reasons. A product may have a short life and may lose popularity after that. Alternatively some new formulation is developed which may be superior to products available in the market. This necessitates comparison of the new product with competitors. In a typical study a product is tried on a group of panelists and some trait is measured before and after use of product. Examples of such traits are skin oiliness or softness or fairness. 
}
\usage{data(Cosmetic1)}
\format{
  A data frame with 48 observations on the following 3 variables.
  \describe{
    \item{\code{Treatment}}{Product code}
    \item{\code{Initial}}{Initial Value of trait}
    \item{\code{Change}}{Change in trait}
  }
}
\details{
Try ANOCOVA. }
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\examples{
data(Cosmetic1)
}
\keyword{datasets}
