\name{Frogfood}
\alias{Frogfood}
\docType{data}
\title{
Study of growth and food preference over age in frogs
}
\description{
When an animal is to be harvested for food it is important to understand its growth pattern and energy requirement. In case of frogs it is of interest to study the relation age X body weight and fit a suitable curve. Further it appears that as age increases, the dietary pattern undergoes change. Hence it is also relevant to study relationship of age with each of 4 components of diet in variables of intake from crabs to total.
}
\usage{data(Frogfood)}
\format{
  A data frame with 7 observations on the following 6 variables.
  \describe{
    \item{\code{Age}}{Age in year group}
    \item{\code{Body_Weight}}{Body weight in grams}
    \item{\code{Intake_Crabs}}{Intake of crabs in grams}
    \item{\code{Intake_Insects}}{Intake of insects in grams}
    \item{\code{Intake_Larvae}}{Intake of larvae in grams}
    \item{\code{Total_intake}}{Total intake in grams}
  }
}
\details{
Explore the use of linear regression, curvilinear regression, and examine their residuals.
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\examples{
data(Frogfood)
}
\keyword{datasets}
