\name{gpk-package}
\alias{gpk-package}
\alias{gpk}
\docType{package}
\title{
100 Data Sets for Statistics Education
}
\description{
Collection of datasets as prepared by Profs. A.P. Gore, S.A. Paranjape, and M.B. Kulkarni of Department of Statistics, Poona University, India. With their permission, this package has been built by me and made available for the benefit of R users. This collection requires a rich class of models and can be a very useful building block for a beginner. 
}
\details{
\tabular{ll}{
Package: \tab gpk\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-12\cr
License: \tab GPL-2 \cr
}
Simply, follow the document files at the website of the data sets. 
}
\author{
Prabhanjan Tattar

Maintainer: Prabhanjan Tattar <prabhanjannt@gmail.com>
}
\keyword{ package }
