\name{Highjump}
\alias{Highjump}
\docType{data}
\title{
Guessing the gold medal score for 2004 Olympics
}
\description{
Olympics are being organized every 4 years for over a century. Performance of the gold medallist generally goes on improving with some cases of reversals. It is of interest to anticipate the results for the next tournament. It is good to remember that the observed values represent extreme cases.
}
\usage{data(Highjump)}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{year}
    \item{\code{Height}}{Height of jump in meters for the champion}
  }
}
\details{
 Regression and time series may be used for the prediction purpose. }
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
