\name{hundredmrun}
\alias{hundredmrun}
\docType{data}
\title{
Guessing the gold medal score for 2004 Olympics
}
\description{
Olympic games are organized every 4 years for over a century. Performance of the gold medallist generally goes on improving with some cases of reversals. It is of interest to anticipate the results for the next tournament. It is good to remember that the observed values represent extreme cases.
}
\usage{data(hundredmrun)}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Year}
    \item{\code{Time.sec.}}{Time in seconds for the champion (male)}
  }
}
\details{
Regression and time series are suggested for this problem. }
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
