\name{IMR}
\alias{IMR}
\docType{data}
\title{
Changes in Infant mortality over last century across countries
}
\description{
Infant mortality is an important indicator of health status of a country. It is of interest to see how different countries have progressed in this aspect over last century, in particular if the relative ranks of countries remain roughly the same.
}
\usage{data(IMR)}
\format{
  A data frame with 8 observations on the following 16 variables.
  \describe{
    \item{\code{Country}}{Name of the Country}
    \item{\code{IMR_1900}}{IMR (per 1000 live births) in 1900}
    \item{\code{IMR_1950}}{IMR (per 1000 live births) in 1950}
    \item{\code{IMR_1985}}{IMR (per 1000 live births) in 1985}
    \item{\code{IMR_1993}}{IMR (per 1000 live births) in 1993}
  }
}
\details{
Experiment with comparison of trend and rank correlation for the data under consideration. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
