\name{IOCSharePrice}
\alias{IOCSharePrice}
\docType{data}
\title{
Modeling share price series of IOC}
\description{
Time series is a sequence ordered in time. Its modeling is of considerable interest to users for purposes of forecasting. 
}
\usage{data(IOCSharePrice)}
\format{
  A data frame with 250 observations on the following 2 variables.
  \describe{
    \item{\code{Date}}{Date }
    \item{\code{Opening_Price}}{Opening price}
  }
}
\details{
Plotting, Trend fitting, Estimating seasonal component, ARIMA modeling, and Residual analysis are some of the powerful tools which may give rich insight into the share prices. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
