\name{Ivoryweight}
\alias{Ivoryweight}
\docType{data}
\title{
Trends in illegal ivory trade
}
\description{
Some people buy products made of ivory (elephant tusks). Hence there is temptation to kill tusker elephants illegally (poaching). There is international agreement that police should arrest such poachers as well as traders in ivory or products made from it. The intention is to reduce illegal killing of elephants so that their populations are preserved. The data give number of pieces and weight of ivory confiscated.
It is of interest to see if there is a change in the ivory trade over the time, both in terms of pieces and in terms of weight. It is also of interest to check the relationship between number of pieces and weight. The analysis may be done for three stages raw, semi worked and worked separately or for total weight.
}
\usage{data(Ivoryweight)}
\format{
  A data frame with 42 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year}
    \item{\code{Ivory}}{The type of ivory}
    \item{\code{Pieces}}{The number of ivory pieces recovered}
    \item{\code{Weight}}{The total weight of the ivory pieces}
  }
}
\details{
Time series, trend analysis, regression analysis need to be attempted for explaining the variation in this data set. }
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
