\name{Lognormalbirds}
\alias{Lognormalbirds}
\docType{data}
\title{
Species abundance distribution
}
\description{
One approach to biodiversity measurement is fitting a distribution to species abundance data. In the present case data on abundance of different species is recorded at two Indian localities in western-ghats (Arvind and Dhoni). These data are said to be 'Frequency of frequency'.
Some ecological considerations suggest that a log-normal distribution should fit the data. S - the number of species observed and area under the curve to the left of 1/2  are used together to estimate the number of 'unseen' species. 
}
\usage{data(Lognormalbirds)}
\format{
  A data frame with 305 observations on the following 3 variables.
  \describe{
    \item{\code{Site}}{The sites}
    \item{\code{Abundance}}{The number of species with abundance r each}
    \item{\code{Species}}{The numberof individual birds sighted }
  }
}
\details{
Report your findings with fitting log-normal distribution and Chi-square test for Goodness of fit.
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
