\name{Lung}
\alias{Lung}
\docType{data}
\title{
Smoking and Lung capacity study
}
\description{
That smoking causes cancer, is widely known. What is not so well known is that even when you escape cancer, chronic smoking can have an adverse effect on ability of lungs. In a clinical trial, a drug for improving lung capacity was administered in two localities to a number of patients. Lung capacity index was measured before and after treatment of one week. The objective is to test efficacy of the drug after discounting for differences between patients and localities. 
}
\usage{data(Lung)}
\format{
  A data frame with 41 observations on the following 8 variables.
  \describe{
    \item{\code{Serial_Number}}{Serial number}
    \item{\code{Locality}}{Locality}
    \item{\code{LCI_Before}}{Lung Capacity index (LCI) before treatment}
    \item{\code{LCI_After}}{Lung Capacity index (LCI) after treatment}
    \item{\code{Age}}{Age}
    \item{\code{Gender}}{Gender}
    \item{\code{Weight}}{Weight}
    \item{\code{Smoking_Index}}{Smoking index (based on intensity and duration of smoking)}
  }
}
\details{
ANOCOVA!
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
