\name{magazine}
\alias{magazine}
\docType{data}
\title{
Time trends in authorship distribution
}
\description{
A monthly magazine in Marathi, wishes to examine trends in the distribution of total number of pages, into different forms of writing. The desire is that role of editorial writing and material from other members of the editorial board should decline. Share of articles by outsiders should increase as also reader's reactions. 
}
\usage{data(magazine)}
\format{
  A data frame with 14 observations on the following 8 variables.
  \describe{
    \item{\code{Year}}{Year}
    \item{\code{Cover_Page}}{Cover page}
    \item{\code{Editor}}{Editor}
    \item{\code{Others_Editorial_Board}}{Others in Editorial board}
    \item{\code{Articles}}{Articles}
    \item{\code{Reprints_Marathi}}{Reprints in marathi}
    \item{\code{Reprints_Other_Languages}}{Reprints in other languages}
    \item{\code{Readers_Reaction}}{Reader's reaction}
  }
}
\details{
Time series and regression may be experimented with. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
