\name{Mammals}
\alias{Mammals}
\docType{data}
\title{
Birth weight and brain size of mammals
}
\description{
Generally larger the animal larger is its brain size. But some species may be exceptional. The question of interest is 'does birth weight predict brain size?' (Data are averages for different mammal species). If there are any exceptional cases, find out which species they are. 
}
\usage{data(Mammals)}
\format{
  A data frame with 99 observations on the following 2 variables.
  \describe{
    \item{\code{Birth_Weight}}{Birth weight}
    \item{\code{Adult_Brain_Weight}}{Adult Brain weight}
  }
}
\details{
Regression model with a pre-cursor of transformation may be effective for this data set. }
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
