\name{Plaque}
\alias{Plaque}
\docType{data}
\title{
Studying effect of toothpaste on plaque accumulation
}
\description{
Accumulation of plaque on teeth is a major cause of dental ill health. Producers of toothpaste often claim that use of their product will reduce the plaque. Data on three products are to be used to check two things (a) does a particular product reduce plaque? (b) Is reduction achieved the same for all products? 
}
\usage{data(Plaque)}
\format{
  A data frame with 60 observations on the following 3 variables.
  \describe{
    \item{\code{Product}}{The product}
    \item{\code{Before}}{Plaque score before treatment }
    \item{\code{After}}{Plaque score AFTER treatment}
  }
}
\details{
Paired t-test and one-way ANOVA may be attempted for this data set. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
