\name{Plastic}
\alias{Plastic}
\docType{data}
\title{
Seasonality in sales of plastic granules
}
\description{
A manufacturer of plastic granules wants to study sales pattern over different months. Comparing different sales officers as well as two plants is also of interest. Any patterns that can suggest suitable action by management are of interest. 
}
\usage{data(Plastic)}
\format{
  A data frame with 1000 observations on the following 4 variables.
  \describe{
    \item{\code{Plant_Code}}{Plant Code 1 or 2}
    \item{\code{Month}}{Month}
    \item{\code{Employee_Code}}{Employee code}
    \item{\code{Quantity}}{Quantity of sales (Metric Ton)}
  }
}
\details{
Try out Histogram, ANOVA, and paired comparisons.
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
