\name{Poliocases}
\alias{Poliocases}
\docType{data}
\title{
The number of polio cases
}
\description{
Poliomyelitis is a crippling disease with dramatically visible impact on the patient. Over the last fifty years the disease has been brought under control by the use of oral vaccine. It is of considerable interest to identify trend, seasonality and other features of data on incidence of polio.
}
\usage{data(Poliocases)}
\format{
  A data frame with 180 observations on the following 3 variables.
  \describe{
    \item{\code{Month}}{Month}
    \item{\code{Year}}{Year}
    \item{\code{Polio_Cases}}{Number of polio cases}
  }
}
\details{
Time series analysis, what else?
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
