\name{Rabbit}
\alias{Rabbit}
\docType{data}
\title{
Relating Foot length to Body mass
}
\description{
Birds of prey such as eagles or falcons catch a prey animal and then take it to a suitable place (nest if young ones are to be fed), remove inedible portion (feathers or bones etc) and then eat the meat. Remaining material is dropped to the ground. If some indigestible material is   swallowed, it is formed into a ball and regurgitated. All such residues dropped down by the predator can provide good clues to its dietary patterns. Rabbits are common prey of eagles. One inedible portion of a rabbit's body is a hind foot. An ornithologist studied the relationship between foot length and total body mass in rabbits. The idea was that by observing the foot length, one might be able to estimate the meat intake.}
\usage{data(Rabbit)}
\format{
  A data frame with 141 observations on the following 2 variables.
  \describe{
    \item{\code{Hind_Foot_Length}}{Rabbit Hind Foot length in inches}
    \item{\code{Body_Weight}}{Body Weight in pounds}
  }
}
\details{
Regression model and analysis of residuals need to be performed on this data set. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}


