\name{RiceWheat}
\alias{RiceWheat}
\docType{data}
\title{
Modeling Rice and Wheat production
}
\description{
At the time of independence India was faced with food shortages. There was dependence on food imports from USA etc. Later, because of green revolution, production went up and the country became self-sufficient. Was this growth in production linear? Was it due to increase in total area under crop or area under irrigation? Is the pattern of growth comparable in wheat and rice?
}
\usage{data(RiceWheat)}
\format{
  A data frame with 106 observations on the following 6 variables.
  \describe{
    \item{\code{Food}}{The type of food}
    \item{\code{Year}}{The year}
    \item{\code{Area}}{Area (million Hectares)}
    \item{\code{Production}}{Production (Million Tons)}
    \item{\code{Yield}}{Yield (Kilogram/Hectare) }
    \item{\code{Irrigated}}{Percentage area for the food type covered by irrigation}
  }
}
\details{
Model this data set with time series analysis and regression. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
