\name{salamander}
\alias{salamander}
\docType{data}
\title{
Habitat preference of salamander
}
\description{
Salamanders are amphibians, a group of animals of special interest because of their sensitivity to environmental changes. A particular species (about 6cm in length) found in California was studied to check the habitat preferred by the animal.
}
\usage{data(salamander)}
\format{
  A data frame with 47 observations on the following 4 variables.
  \describe{
    \item{\code{Site}}{Site number}
    \item{\code{Salamander}}{Salamander count in a sample plot of 7meter by 7 meter area}
    \item{\code{Coverage}}{Extent of canopy cover in the forest}
    \item{\code{Forest_Age}}{Forest age}
  }
}
\details{
Anybody for Poisson regression?
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
