\name{TeethNormal}
\alias{TeethNormal}
\docType{data}
\title{
Modeling indicators of dental health 
}
\description{
There are three widely accepted indicators of dental health. They are gingival score, bleeding (gums) score and plaque score. It is always of interest to check effect of toothpaste on these variables. To decide which methods of statistical analysis would be appropriate, it is relevant to test normality of these measurements.
}
\usage{data(TeethNormal)}
\format{
  A data frame with 69 observations on the following 3 variables.
  \describe{
    \item{\code{Gingival}}{Gingival score}
    \item{\code{Bleeding}}{Bleeding (of gums) score}
    \item{\code{Plaque}}{Plaque score}
  }
}
\details{
Effective use of histogram, P-P plot, and goodness of fit may provide the answers. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
