\name{Tiger7}
\alias{Tiger7}
\docType{data}
\title{
Identification of individual tigers from pugmarks
}
\description{
Refer Tiger7.doc for precise explaination of the data set. 
}
\usage{data(Tiger7)}
\format{
  A data frame with 78 observations on the following 7 variables.
  \describe{
    \item{\code{P_TC1_Dist}}{Distance between Pad center and Toe 1 center}
    \item{\code{P_TC2_Dist}}{Distance between Pad center and Toe 2 center}
    \item{\code{P_TC3_Dist}}{Distance between Pad center and Toe 3 center}
    \item{\code{P_TC4_Dist}}{Distance between Pad center and Toe 4 center}
    \item{\code{TC1_TC2_Dist}}{Distance between Toe1 center and Toe 2 center}
    \item{\code{TC2_TC3_Dist}}{Distance between Toe2 center and Toe 3 center}
    \item{\code{TC3_TC4_Dist}}{Distance between Toe3 center and Toe 4 center}
  }
}
\details{
Cluster analysis is recommended for gaining an insight into this problem. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
