\name{TigerIdentity}
\alias{TigerIdentity}
\docType{data}
\title{
Tiger census using scat samples
}
\description{
Tiger census is a long debated issue in wildlife conservation. Since direct count is not possible any indirect method is based on similarities/ discrepancies in sample evidence. Two methods, which have received attention, are (i) pugmark counting and (ii) camera trap method. In pugmark method, if two pugmarks appear similar in shape and size, they are supposed to represent same animal. In Camera trap, if two photographs show similar stripe pattern, they represent same tiger. It now appears that a third method is possible. It is based on parasite composition of scat samples. If the composition is similar, most likely the two samples come from same animal. 
}
\usage{data(TigerIdentity)}
\format{
  A data frame with 55 observations on the following 33 variables.
  \describe{
    \item{\code{ID}}{Sample ID}
    \item{\code{Type.1}}{Abundance of various pathogen types, 1 to 32, in the sample}
    \item{\code{Type.2}}{}
    \item{\code{Type.3}}{}
    \item{\code{Type.4}}{}
    \item{\code{Type.5}}{}
    \item{\code{Type.6}}{}
    \item{\code{Type.7}}{}
    \item{\code{Type.8}}{}
    \item{\code{Type.9}}{}
    \item{\code{Type.10}}{}
    \item{\code{Type.11}}{}
    \item{\code{Type.12}}{}
    \item{\code{Type.13}}{}
    \item{\code{Type.14}}{}
    \item{\code{Type.15}}{}
    \item{\code{Type.16}}{}
    \item{\code{Type.17}}{}
    \item{\code{Type.18}}{}
    \item{\code{Type.19}}{}
    \item{\code{Type.20}}{}
    \item{\code{Type.21}}{}
    \item{\code{Type.22}}{}
    \item{\code{Type.23}}{}
    \item{\code{Type.24}}{}
    \item{\code{Type.25}}{}
    \item{\code{Type.26}}{}
    \item{\code{Type.27}}{}
    \item{\code{Type.28}}{}
    \item{\code{Type.29}}{}
    \item{\code{Type.30}}{}
    \item{\code{Type.31}}{}
    \item{\code{Type.32}}{}
  }
}
\details{
Cluster analysis is again recommended. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
