\name{Timber}
\alias{Timber}
\docType{data}
\title{
Genetic and environmental components of tree characteristics
}
\description{
Refer Timber.doc from the web link given below for a detailed description of the problem. 
}
\usage{data(Timber)}
\format{
  A data frame with 224 observations on the following 10 variables.
  \describe{
    \item{\code{Locality}}{Locality}
    \item{\code{Year}}{Year of experiment (two years, 2000 and 2002)}
    \item{\code{Replicate}}{Replicate}
    \item{\code{Subculture}}{Subculture}
    \item{\code{Elongation}}{Percent of Elongation}
    \item{\code{Multiples}}{Percent of Multiples}
    \item{\code{Rooting}}{Percent of rooting}
    \item{\code{Germination}}{Percent of germination}
    \item{\code{Seed_Length}}{Seed length in cm}
    \item{\code{Seed_Width}}{Seed width in cm}
  }
}
\details{
 ANOVA, Variance component analysis, and PCA are recommended for the data set. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
