\name{widowbird}
\alias{widowbird}
\docType{data}
\title{
Mate selection by females
}
\description{
Refer the document titled widowbirdexpt.doc obtainable from the link given below. 
}
\usage{data(widowbird)}
\format{
  A data frame with 36 observations on the following 5 variables.
  \describe{
    \item{\code{New_Tests}}{Number of new nests}
    \item{\code{Treatment_Group}}{Treatment group}
    \item{\code{Male}}{ID of the male bird}
    \item{\code{Tail}}{Tail length cm.}
    \item{\code{Prev}}{Previous nest count}
  }
}
\details{
ANOCOVA may be tried out here. 
}
\source{
http://ces.iisc.ernet.in/hpg/nvjoshi/statspunedatabook/databook.html
}
\keyword{datasets}
